added html report generation for qq-compare. Created utils.py inside plot folder to move aux plotting functions
