+++ src/qibocal/plots/heatmaps.py
@@ -5,17 +5,7 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
 from qibocal.data import DataUnits
-
-
-def get_data_subfolders(folder):
-    # iterate over multiple data folders
-    subfolders = []
-    for file in os.listdir(folder):
-        d = os.path.join(folder, file)
-        if os.path.isdir(d):
-            subfolders.append(os.path.basename(d))
-
-    return subfolders[::-1]
+from qibocal.plots.utils import get_data_subfolders
 
 
 # Resonator spectroscopy flux
