+++ src/qibocal/plots/utils.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+import os.path
+
+def get_data_subfolders(folder):
+    # iterate over multiple data folders
+    subfolders = []
+    for file in os.listdir(folder):
+        d = os.path.join(folder, file)
+        if os.path.isdir(d):
+            subfolders.append(os.path.basename(d))
+
+    return subfolders[::-1]
